+++ milkman-jdbc/src/main/java/milkman/plugin/jdbc/editor/JdbcRequestEditor.java
@@ -6,16 +6,20 @@ import javafx.scene.control.TextField;
 import lombok.SneakyThrows;
 import milkman.domain.RequestContainer;
 import milkman.plugin.jdbc.domain.JdbcRequestContainer;
+import milkman.ui.components.AutoCompleter;
+import milkman.ui.plugin.AutoCompletionAware;
 import milkman.ui.plugin.RequestTypeEditor;
 import milkman.utils.fxml.FxmlUtil;
 import milkman.utils.fxml.GenericBinding;
 
-public class JdbcRequestEditor implements RequestTypeEditor {
+public class JdbcRequestEditor implements RequestTypeEditor, AutoCompletionAware {
 
 	
 	@FXML TextField jdbcUrl;
 	
 	private GenericBinding<JdbcRequestContainer, String> urlBinding = GenericBinding.of(JdbcRequestContainer::getJdbcUrl, JdbcRequestContainer::setJdbcUrl);
+
+	private AutoCompleter completer;
 	
 	
 	@Override
@@ -34,7 +38,13 @@ public class JdbcRequestEditor implements RequestTypeEditor {
 		
 		urlBinding.bindTo(jdbcUrl.textProperty(), restRequest);
 		urlBinding.addListener(s -> request.setDirty(true));
+		completer.attachVariableCompletionTo(jdbcUrl);
 		
+	}
+
+	@Override
+	public void setAutoCompleter(AutoCompleter completer) {
+		this.completer = completer;
 	}