+++ milkman-rest/src/main/java/milkman/ui/plugin/rest/RequestHeaderTabController.java
@@ -1,22 +1,32 @@
 package milkman.ui.plugin.rest;
 
+import static milkman.utils.FunctionalUtils.run;
+
+import java.io.InputStream;
+import java.util.List;
+import java.util.UUID;
+
+import org.apache.commons.io.IOUtils;
+import org.apache.commons.lang3.ClassPathUtils;
+
 import javafx.scene.control.Tab;
 import lombok.SneakyThrows;
-import milkman.domain.RequestAspect;
 import milkman.domain.RequestContainer;
+import milkman.ui.components.AutoCompleter;
 import milkman.ui.components.JfxTableEditor;
+import milkman.ui.plugin.AutoCompletionAware;
 import milkman.ui.plugin.RequestAspectEditor;
 import milkman.ui.plugin.rest.domain.HeaderEntry;
 import milkman.ui.plugin.rest.domain.RestHeaderAspect;
 import milkman.utils.fxml.FxmlUtil;
 
-import static milkman.utils.FunctionalUtils.*;
-
-import java.util.UUID;
-
-public class RequestHeaderTabController implements RequestAspectEditor {
+public class RequestHeaderTabController implements RequestAspectEditor, AutoCompletionAware {
 
 
+	private AutoCompleter completer;
+
+	private static List<String> headers;
+	
 	@Override
 	@SneakyThrows
 	public Tab getRoot(RequestContainer request) {
@@ -27,8 +37,15 @@ public class RequestHeaderTabController implements RequestAspectEditor {
 			return new HeaderEntry(UUID.randomUUID().toString(), "", "", true);
 		});
 		editor.addCheckboxColumn("Enabled", HeaderEntry::isEnabled, run(HeaderEntry::setEnabled).andThen(() -> headers.setDirty(true)));
-		editor.addColumn("Name", HeaderEntry::getName, run(HeaderEntry::setName).andThen(() -> headers.setDirty(true)));
-		editor.addColumn("Value", HeaderEntry::getValue,run(HeaderEntry::setValue).andThen(() -> headers.setDirty(true)));
+		List<String> headerList = getHeaders();
+		editor.addColumn("Name",
+				HeaderEntry::getName, 
+				run(HeaderEntry::setName).andThen(() -> headers.setDirty(true)),
+				tf -> completer.attachStaticCompletionTo(tf, headerList));
+		editor.addColumn("Value", 
+				HeaderEntry::getValue,
+				run(HeaderEntry::setValue).andThen(() -> headers.setDirty(true)),
+				tf -> completer.attachVariableCompletionTo(tf));
 		editor.addDeleteColumn("Delete", () -> headers.setDirty(true));
 		
 		editor.setRowToStringConverter(this::headerToString);
@@ -47,6 +64,20 @@ public class RequestHeaderTabController implements RequestAspectEditor {
 		if (!header.isEnabled())
 			prefix = "//";
 		return prefix + header.getName() + ": " + header.getValue();
+	}
+
+	@Override
+	public void setAutoCompleter(AutoCompleter completer) {
+		this.completer = completer;
+	}
+	
+	@SneakyThrows
+	private static List<String> getHeaders() {
+		if (headers == null) {
+			InputStream inputStream = RequestHeaderTabController.class.getResourceAsStream("/headers.txt");
+			headers = IOUtils.readLines(inputStream);
+		}
+		return headers;
 	}
 
 }