+++ milkman/src/main/java/milkman/ui/main/RequestComponent.java
@@ -60,6 +60,7 @@ public class RequestComponent {
 		plugins.loadRequestAspectPlugins().stream()
 		.flatMap(p -> p.getRequestTabs().stream())
 		.forEach(tabController -> {
+			plugins.wireUp(tabController);
 			if (tabController.canHandleAspect(request)) {
 				if (tabController instanceof ContentTypeAwareEditor) {
 					((ContentTypeAwareEditor) tabController).setContentTypePlugins(plugins.loadContentTypePlugins());